requirements.txt
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.3.0
umap-learn>=0.5.4
# Pinecone SDK (rename from pinecone-client)
pinecone==5.3.0
# Frontends
streamlit==1.37.1
python-dotenv>=1.0.0
